requirements.txt
openpyxl==3.1.5
pandas==2.2.3
psycopg2_binary==2.9.10
pydantic_settings==2.9.1
pydantic==2.9.1
sqlalchemy==2.0.41
uvicorn==0.34.2
fastapi==0.115.12
starlette==0.46.2
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
python-dotenv==1.0.0
fastapi-cors==0.0.6
fastapi-cache2==0.2.2